requirements.txt
streamlit==1.31.1
opencv-python-headless==4.8.1.78
mediapipe==0.10.9
numpy==1.26.3
Pillow==10.2.0